refactor: move S3ObjectStore to s3_objectstore.go
